Classify all requirements in new YAML file.
YAML because I want this to human-writeable but machine-readable.

This classifies all requirements from LDM-556 by whether we've already
implemented the functionality or not (and whether it's even worth
doing so in the future), along with some notes and in some cases
high-level text "scripts" for how we might verify them.  Those fell by
the wayside for later requirements after it became clear that we might
want to claim a lot of the more trivial requirements based on
daf_butler unit tests, regular HSC/DC2 reprocessing, DP0.2, or the
first Gen3 acceptance milestone testing.